remove unnecesary type cast of returned 0

+++ 0x07-pointers_arrays_strings/3-strspn.c
@@ -26,5 +26,5 @@ unsigned int _strspn(char *s, char *accept)
 		}
 	}
 
-	return ((unsigned int)0);
+	return (0);
 }
